Introduce new HTTP Query API ()
This change adds a revamped version of the HTTP Query API, formerly known as HTTP Transaction API. It is implemented as an additional Jakarta RS Resource under the name `QueryResource`. The new resource is mapped as `/{databaseName}/query/v2`.

In this first instalment, the new resource must be explicitly enabled via `server.http_enabled_modules=QUERY_API_ENDPOINTS`. As that property is not additive, any other options such as delivery of Browser files, or the old API must be enumerated as well if the should be available, too (i.e. like this `server.http_enabled_modules=BROWSER,QUERY_API_ENDPOINTS,TRANSACTIONAL_ENDPOINTS`).

The new resources uses a special instance of the Neo4j Java Driver connected via a local, in-process Netty channel, thus all interaction goes through the driver. The API is therefor able to run in a cluster in case server side routing is enabled via `dbms.routing.enabled=true` and the necessary listening and advertised addresses.

The format is heavily inspired by the original PoC of a standalone HTTP module for Neo4j, but still subject to change. The type representation is different from JOLT and much more aligned with other APIs aiming at transporting types over JSON. This should make parsing the format for downstream clients a lot easier.

Co-authored-by: Michael Simons <michael.simons@neo4j.com>