Reinventing degree store again ()
This PR replaces "old" VersionedDegreesStore with new VersionedCachedDegreeStore 🎉

New store uses cache, and all read or modified degrees are kept in cache for fast operations. Periodically (during checkpoint atm) the cache is flushed to the store which is multi root tree.

Multi root tree is used here to provide an easy way to cleanup/overwrite all deltas related to the single degree when flushing the cache.