Change QueryGraphPredicates.externalLabelInfo
Previously, when encountering a node in cardinality estimation
that was an argument, we would use these labels for its cardinality estimation:
All the labels from previous query parts, excluding the ones that are also in the current query graph.
This logic was flawed.

Now, for such an argument node, we use these labels for its cardinality estimation:
All the labels from previous query parts.

Examples:
* `MATCH (a:FOO) ... MATCH (a)-[:TYPE]->(b:BAR)`
  We correctly use `:FOO` to estimate `a` in the 2nd query graph. This was also done previously.
* `MATCH (a:FOO) ... MATCH (a:FOO)-[:TYPE]->(b:BAR)`
  We correctly use `:FOO` to estimate `a` in the 2nd query graph. This was _not_ done previously.
* `MATCH (a) ... MATCH (a:FOO)-[:TYPE]->(b:BAR)`
  We correctly skip using `:FOO` to estimate `a` in the 2nd query graph.
  This was also done previously.