Less reserve/unreserve dance when looking for IDs to cache
Basically maintenance() will interfere less with normal operations.
This is only a problem for multi-ID generators where maintenance will
see that the cache isn't full and there are IDs to likely find in the
tree, but it may turn out that all or most IDs found are of sizes that
the cache is already full of and so there would be the pessimistic
reservation of all those IDs, only to discover right afterwards that
they wouldn't fit and they would be marked as uncached.