Label inference for undirected rels
Inferring node labels from types of undirected relationships (in combination with graph statistics).
When the cardinality of ()-[:S]->() is the same as (:A)-[:S]->(), then label A can be inferred on the start node.
When the cardinality of ()-[:S]->() is the same as ()-[:S]->(:A), then label A can be inferred on the end node.
When a label can be inferred on both the start and end nodes, then it can be inferred on both nodes of an undirected relationship.
When a label can be inferred on the start node or the end node but not on both, then the label cannot be inferred on any of the nodes of an undirected relationship.