5 seed syncing ()
Seed syncing
In broad strokes this PR consists of 5 major changes.

Firstly, we've made adjustments to the existing seed validation component to make it work nicely with seed syncing. The overall logic is the same, although we replaced the rules (ValidationRule) with different validators (SeedValidator). Previously we'd use rules to get different behaviours from the validation. Now we'll use validators to do so. So in the future if we want to add another case of validation, we can add another implementation of SeedValidator. Also the strategies (RemoteSeedValidationStrategy) have been removed. Now it's up to the SeedValidator to decide which outcomes are ok and not.


Furthermore, the logic in RemoteSeedValidator has been split up between SeedSyncingClient and the SeedValidator. The actual remote call will be executed with the SeedSyncingClient. Then it's up to the SeedValidator to use the responses and send it through the SeedComparator. Which in turn returns a SeedComparisonOutcome. It tells whether the local seed is IDENTICAL/BEHIND/INCOMPATIBLE to the remote seed.


So a typical SeedValidator will call the SeedSyncingClient with the remotes that are important to it. Then it'll compare its LocalSeedContext to the remotes with the SeedComparator. Then depending on what criteria it has, it'll decide the status of the local seed. The SeedValidator will return SeedValidatorOutcome. The various validators often differ in which remotes it should compare to, and what criteria it has. Also, the validators decide what can be done with the seed. For example, there are validators that don't support seed updating. Those will never return outcomes that trigger the updater.

Add a SeedSyncingController which takes the outcome from the SeedValidator and depending on the outcome calls the various seed syncing components. For example it can create a temporary database, trigger the seed updater, throw seed or exit.

Add SeedUpdater which is responsible for catching up the offline seed. Also add a special CatchupAddressProvider which contact all other members if the database is offline. If it's online, we'll use the normal address provider. The offline address provider is a best effort, round robin thing. It'll return the addresses one by one, until we found one that's able to serve our catchup request. We've deliberately put the responsibility on the server to decide whether it's ok to serve requests or not. The client, will simply try until it finds anyone. This can surely be optimised in the future.

Add SeedSyncingTemporaryDatabaseFactory for keeping track of the temporary database created for the seed syncing controller. Due to the limitation in the catchup API which requires the database to be online, we have to create a fake temporary database over the real store files, in order to catchup the seeds in seed syncing. Also we need to connect it to the Catchup server so that in case there's a temporary database, the data can be fetched from there instead of the real kernel database.

Add SeedSyncingComponent to the system graph components. This is used to prevent members from trying seed updating before everyone supports it. Once everyone has converged to the SEED_UPDATING_SUPPORTED version, seed updating will be turned on. Until then all members will just execute seed validation.

Additional changes:

Add ValidatedStoreId. This class is used in SeedSyncingLifecycle to store the validated storeId. This storeId is later used in the ClusterStorageEngineFactorySupplier to retrieve the SEF that the database should use. We must explicitly set ValidatedStoreId, but it may be UNKNOWN in case no-one has a store. If the storeId is UNKNOWN we read the SEF from the config.


We're happy to keep the old implementation if the reviewer thinks it's too many logical changes. The reason we changed this is because the old one sent down a large component with a lot of SEF logic into seed validation. We figured a smaller class would be more controlled and safer. Also there was nothing asserting that we had actually used the validated storeId to set the SEF. If we in any case by mistake would forget to set the validated storeId, we'd use the SEF in config which is not guaranteed to be the correct one for seeded databases.

In catchup, always read the SEF from the downloaded storeId. In incremental catchup there's no storeId downloaded. In that case, we can either read the SEF from the kernel database (if it's started), otherwise, we read from disk.


Co-authored-by: Anna Sjerling <anna.sjerling@neotechnology.com>