Block format importer splits up rel input into partitions for scalability
Partitions are basically ranges of node IDs where relationships ends up
in one out of all ranges depending on which is their lowest node.
Then when importing, going from range to range, relationships are imported
for the sides that are covered in the current range so that the page cache
access is much more controlled.