Fix deadlock in parallel checking of MultiRoot tree, and enable it again
The problem:
Data trees are checked in batches and they all submit a job to the executor. For each batch
each of the data trees do consistency check - that check is parallelized as well and each
child of that root becomes a separate task. Those tasks are submitted to the same executor.
After having registered all its children tasks the thread used to await those futures.
The inner tasks just do their job and finish, but if there are enough of the ‘outer’
jobs waiting for all futures of their children to finish every thread could end up hanging
in awaitAllFutures.

The fix tries to solve this by not having any of the data trees wait for
the child tasks, those tasks are instead waited on at the end of the
outmost check.