Micro-optimize conflicts Map key.
We used Sets of fixed size 2 as key in the conflicts Map.
Creating these Sets turns out to take quite some time, if there are many conflicts.

A hand-crafted class takes virtually no time to create.
Downside is that we have to manually implement hashCode, as creating a Set to calculate its hashCode
would defeat the whole purpose of the optimization.