Schedule CompressPlanIDs twice
EagerRewriter can accumulate quite some memory,
so it is better to release as much as possible before.
In this case, compressing plan IDs allows releasing solved of
unused plans.

---

This also makes the dependency between SortPredicatesBySelectivity and InsertCachedProperties
explicit as a precondition (and not just an invalidated condition), because otherwise
SortPredicatesBySelectivity would get scheduled twice.