Changes to some cases of truncating logs in recovery
There was a case in recovery where corrupted transaction logs could be truncated even without the fail_on_corrupted_logs
setting disabled. If the first byte on a transaction boundary was 0 the reader would just think it reached the end of the file
and say there were no more transactions and then truncate would happen. Now it detects this corruption and reports it
so the user can decide on how to proceed.

Before a corrupt checkpoint file would only be truncated if there were transactions to recover as well.
Now corruption in the checkpoint log file will be truncated away.
This replaces the functionality that was rotating to the next file if
there was unreadable bytes in the checkpoint log. It should always be
truncated instead.