Fix bug in query obfuscation
This commit fix a bug in query obfuscation when pre-parser options are present.
The pre-parser part of a query is not included in query cache keys.
For this reason all input positions that comes from a cache needs to be offset.
The query obfuscator relies on input postions to function.
This commit makes sure obfuscation is offset correctly with regards
to pre-parser options.

Fix issues like this for example:

```
> return 1;
Query log: return ******

> cypher replan=skip return 1;
Query log: cypher ******eplan=skip return 1
```