Remove FailedIndexProxyFactory
The purpose of this is to clean-up the Index API a little bit, by
removing old and unecessary stuff. The FailedIndexProxyFactory interface
and its single implementation are not needed at all, because when
FlippableIndexProxy.flip() fails, the thrown
ExceptionDuringFlipKernelException will be caught by the try-catch on
MultipleIndexPopulator.flipAfterStoreScan(), which is the only caller
for this code path.

When that happens, it will call MultipleIndexPopulator.cancel(), which
already calls population.cancel(indexPopulationFailure) that create a
FailedIndexProxy and flip the index to it.

In effect we were setting the delegate of FlippableIndexProxy twice to
two new instances of FailedIndexProxy on the failure path!