5.x Fail transaction if kernel commit fails ()
Track exceptions from InternalTransactionCommitProcess and use them in onClose callback so that we can notify the user about what went wrong. Previously we'd panic, and depending on where the transaction failed in the commit process, the tx would either hang until interrupted (?) or come back as successful. Now we'll still panic, but the transaction will always come back with the exception that caused the panic.

A local commit failure will have "UnknownError" status with a message saying that the tx failed locally but may have succeeded somewhere else.