Make TxStateTest test what it claimed
shouldVisitDeletedNodesAfterDeletedRelationships was testing
that visiting nodes happened after visiting created rels but
claimed to test visiting deleted nodes after deleted rels.
Changed since created and deleted rels are visited by the same method
nowadays, so it doesn't matter, and the test was flaky since it is not allowed
to delete a node that has any relationships left.