Changing the place where Kernel transactions are wrapped into SPD Kernel transactions
It is currenlty done in Query router. That does not work for at least the following cases:
  - CALL IN TRANSACTIONS
  - Core API
  - procedures starting new transactions

Also having unmodified kernel transactions around in SPD databases poses a danger that they will resurface in unexpected places.
This change wraps SPD transactions immediately as they are created, so no code in the DBMS has a reference to an unmodified transaction.