Introduce forEach method on ProcedureHolder ()
..to avoid an unnecessary call to `idOf(QualifiedName)` that will cause more problems when we begin to include `CypherScope`s in the holder keys.

Since we intend to add CypherScope to the keys, we will naturally extend `idOf` to `idOf(QualifiedName, CypherScope)`. However, when these methods are used (during the creation of the securityContext) we are not necessarily aware of which CypherScope to use.

All in all, this change allows us to clean up the code a little bit, leaving the overall class looking slightly tidier.