Bug fix for SubtractionNodeLabelIndexCursor in runtime
MATCH (n:A&B&!C&!D) should match nodes that have the label A and label B and not the label C and not the label D.

A node x with label A,B,C should not be matched. With the intersection scan for the negative labels, node x would have been matched, because node x will appear in the intersection of the positive label scans (A and B), but not in the intersection of the negative label scans (C and D).

Therefore, an intersection scan is needed for the positive labels, but a union scan is needed for the negative labels.