Do not check OrderedIndexPlansUseCachedProperties after NameDeduplication
If we check OrderedIndexPlansUseCachedProperties after NameDeduplication,
it can false-positively complain because variables now share the same name.
We should therefore not check OrderedIndexPlansUseCachedProperties after NameDeduplication.

Also, let `If` not propagate the postConditions of inner.
They cannot be checked in general if the inner transformer is never run.

After having implemented all the changes, I saw that the fix for `If`
alone would actually also solve the added test in this commit.
But only because in `org.neo4j.cypher.internal.compiler.phases.CompilationPhases.planPipeLine`
we wrap the whole `orderedPlanPipelineSteps` in an `If`, and thus do
not check it's postConditions later with this change.