Apply updates for eventually consistent fulltext indexes in parallel
A fulltext index, if setting
dbms.index.fulltext.eventually_consistent=true will not be updated
by the thread applying the transaction that results in changes to it.
Rather the transaction thread will queue the updates for a dedicated
background thread (per dbms) to pick up and apply.
Previously this was always a single background thread and so for
high fulltext index update pressure (e.g. when having many/large
fulltext indexes) this would instead become a bottleneck and
transactions that wanted to queue updates would wait for queue
to become smaller.

This commit adds configurability of having multiple of these
background threads taking from this queue and apply fulltext
index updates. To take things even further, the refreshing
of fulltext indexes will be done with another thread pool
with some configurable interval as to (when eventually consistent):
- take away refresh from main index apply execute path
- generally reduce the number of refreshes for fulltext indexes