Allow procedures to be stored by CypherScope ()
.. by adding a layer of indirection in ProcedureHolder.

Instead of directly mapping to the store item, we now include a
secondary mapping from `CypherScope`->`store item`. Since 
the number of CypherScope's is expected to be small, we use 
an int array to store the mapping.

This extension of ProcedureHolder, also fulfills an expectation of
TokenIdHolder, that if the procedure at point (name, {scopes}) is 
added then we expect that `ProcedureHolder.all()` enumerates this 
item **only** once.