add index setting validators
The main interface that will be passed around is the
VectorIndexSettingsValidator.

This interface has methods to produce VectorIndexConfig objects.

trustIsValidToVectorIndexConfig is intendend for reading preexisting
IndexConfigs, they are treated as truth as they should have been
already validated by some method in its creation.

validateToVectorIndexConfig is the creation path that does the
validation.

validate is the primary checker creating the records, exposed for
Cypher's usage later.

The methods provided that take the validation records are to avoid
having to regenerate the validation records if already done for other
explicitly checking, such as potentially in cypher's validation.