add validator instances to vector index version
Here we separate the different validations by both vector index version
and the kernel version.

As you need to know the vector index version, and the validators require
information _from_ it, it makes sense to include them as part of the
version.

The abstraction left is the map of kernel version to the validation.
As we add more settings, or modify the accepted range of values (even in
an addative way), it's better than they are explicitly versioned.

There is an implied range of validity _between_ map entries; thus having
them in a sorted map gives us access to this without having to
explicitly have a range that we need to manually update and ensure no
overlaps. They are reverse ordered in the map such that the newest
validation is first, thus we only need to check if the provided kernel
version is at least that version to aquire the validator.

If not specified, we use the latest.