add index config validation wrapper
The vector index requires some validation on the index config; currently
that is handled in VectorUtils, and occurs each time the value is used.
Such validations should be versioned, and thus it is easier to extract
that functionality into a nicer more scalable interface.

Other indexes such as POINT and FULLTEXT also take options, which could
be also be validated more than their inputs already are.

It would be too much work currently to add such concepts directly in the
IndexConfig, thus for now will create this wrapper and increase its
support for the other indexes. Perhaps in time it could be refactored to
replace the current IndexConfig usage.