Report when system database is panicked versus otherwise unavailable ()
When a server panics, that triggers the system database to panic, but asynchronously. The database panicking makes it shut down. If the system database has recently changed, the TopologyGraphDbmsModelSnapshotTracker might be trying to produce a new snapshot. If it gets an unavailable exception, it assumes the problem is transient and retries indefinitely. But a panicked system database is permanent.

This could lead to panicked servers hanging in shutdown forever, as seen in the ServerPanicDiagnosticsIT.

The fix is to throw a new SystemDatabasePanickedException sub-exception when the database has panicked. Callers can differentiate based on that if they want to, and the tracker will just give up at that point, unblocking shutdown.