fix(OrLeafPlanner): Should pass the original query graph when updating solveds ()
The OrLeafPlanner was identifying and extracting common predicates across disjunctions. It would then use a new computed query graph to build the union plans. However, as a result of rewriting the predicates the selector in subsequent steps would consider the original predicates as "missing" and add a filter step. For complex/large label expressions this would then fail to verify since the boolean expression generated is different from the original query and the CNF normalizer doesn't apply beyond a certain limit. This commit ensures that we will only pick out common predicates if they were part of the relatedPredicates and not within the original disjunction itself. It was also build the remaining solved queries from the disjunction if possible.