feat: Move neo4j-genai-plugin to Community edition. ()
This commit makes the `neo4j-genai-plugin` part of the community edition. For this to work, several things needed to change:

* The plugin itself must not create `new` instances of any monitory, but need to use `Monitors` facility instead
* The plugin itself must not depend on anything from the enterprise edition, both for technical reasons (circular dependencies) but of course due to licensing questions
* The metrics system on the other hand is not quite ready to receive metrics from components that might or might not be loaded

The latter could have be mitigated to some extend via a third module with shared interfaces for both the plugins monitor and the provider, but that would add even more complexity in deployment to the plugin.

Instead the `GenAIMetrics` have been rewritten in such a way that it checks reflectively during lifecycle `start` whether the plugin classes are available or not. Checks have been implemented as lazy as possible, so it won't add much startup overhead.
If the plugin is available, the listener facilities of `Monitors` are then used to update the counters, when the monitor is called.  To avoid any compile time dependency from the metrics to the plugin, the listener interface is provided as dynamic proxy. 
While the plugin could have been a `provided` dependency of enterprise metrics allowing to avoid the proxy, it would still have been safeguarded against the possible runtime absence of the plugins classes, so we went with the proxy.