feat: rewrite trails using internal node variable to varexpands ()
Expands the trail to var expand rewriter to handle trails that have a predicate on one of the inner node variables.
The rewriter will replace them with either startNode() or endNode() of the corresponding relationship variable. Since this is only predictable on directed relationship queries, the scope of this rewrite is limited just to directed queries.

Also, to actually replace the variable with startNode and endNode, a new RelationExpressionRewriter is introduced. The logic used in the rewriter is effectively extracted from the util ConvertToNfa