Refactor enterprise constraints enforcer
They will do more than only property enforcement with the new constraints,
so we need to rename and reorganise it:
* Move the enforcing logic into [Node|Relationship]Enforcer and
EnterpriseConstraintEnforcer becomes a very simple delegation-like
decorator and we can apply the changes required for new constraints
inside their specific enforcers only.
* ConstraintEnforces only need TokenNameLookup, instead of keeping
the whole StorageReader object, we can pass and keep only the
TokenNameLookup for that.

We also improve the internal documentation of nested maps in
ConstraintEnforcers, since node and relationship enforcers
have a very similar Map<Int, Map<Int, Stuff>> object
that is slightly different on how it is built:
* Nodes version is PropertyKey, then Label.
* Relationship is RelationshipType, then PropertyKey.