Introducing Bolt 5.6 with GqlStatusObject ()
To comply with GQL, the notification and error objects must be updated to include a `GQLSTATUS` code and status description.

Therefore, the current structure of these objects will need to be updated.

While we need to update the format of the errors and notifications to adhere to the GQL standard, we must also maintain the old style for backward compatibility.

The protocol 5.5 was skipped since their specification missed the `description` field and this field is needed for backwards compatibility. Since drivers already implemented Bolt 5.5, Bolt Server should skip the bugged version.

## Bolt 5.6

This new version of Bolt will drop the `notifications` field in favor to the GQL complaint `statuses`field. This new field is a list of dictionary which is define like:

```
statuses::List<Dictionary>( // note this is minified from the API object of GqlStatusObject, this is to save some bytes.
	gql_status::String,
	status_description::String,
	description::String // for backward compatibility
	neo4j_code::String // for backward compatibility
	title::String // for backward compatibility
	diagnostic_record::Dictionary(
		OPERATION::String
		OPERATION_CODE::String
		CURRENT_SCHEMA::String
		_status_parameters::Dictionary
		_severity::String // this is notificationSeverity renamed
		_classification::String // this is notificationCategory renamed
		_position::Dictionary(
			offset::int
			line::int
			column::int
		)
	)
)
```

 Fields on the `diagnostic_record` or the entire object should be omitted if filled with default values.This defaults are versioned along with the protocol version, so newer protocol version can define different defaults.

* Bolt 5.6 Default Diagnostic Record:

```json
{
	"OPERATION": "",
	"OPERATION_CODE": "0",
	"CURRENT_SCHEMA": "/"
}
```

Empty objects are also omitted for saving bandwidth. 

The notification configuration will also be adjusted in this version, the property `notifications_disabled_categories` is renamed to `notifications_disabled_classifications` on `HELLO`, `BEGIN` and `RUN`.

## Design and Implementation notes

### Adding GQL statuses to metadata

For handling the new metadata fields, the `MetadataHandler` supports receiving in `statuses` and the `DefaultMetadataHandler` handles notifications as `statuses` using the Bolt 5.6 specs, including default values. Protocols version between `4.4`  and `5.4` are using the `MetadataHandlerV44` and older protocols (< 4.4) use `MetadataHandlerV40`. `AbstractLegacyMetadataHandler`provides a base implementation for the legacies handlers with support for legacy notifications.  

Extension methods were added to enable protocols that don't support `db` on routing tables or `contains-updates` and `contains-system-updates`to extends with the desirable behaviour, to do not add theses fields.

The `MetadataHandler::onNotifications` was changed to include `Iterable<GqlStatusObject> statuses` as parameter. Older implementations ignore this parameter. The default implementation use it fill the `statuses` field on the metadata. `title` and `neo4j_code`are only field when status are instance of `Notification`and values are not null. 

Empty positions are removed from the diagnostic record since this value means no value.

### Renaming configuration fields

For renaming the configuration field on `HELLO`, `BEGIN` and `RUN`, the `NotificationsConfigMetadataReader`methods were segmented with a new read legacy methods for supporting the `notifications_disabled_categories` while the default methods are using the new configuration property, `notifications_disabled_classifications`.

Legacy implementations were added, `HelloMessageDecoderV53`, `BeginMessageDecoderV52` and `RunMessageDecoderV52`.

### Fix `OPERATION_CODE` in `DiagnosticRecord`

The value should be a `string` instead of a `integer` as specified in the technical document and drivers ADR. 

### Fabric Support

The fabric support to `GqlStatusObject`was done by extending the `ResultSummaryWrapper`class. `GqlStatusObjectWrapper`was added for status which aren't notifications and `GqlStatusObjectNotificationWrapper`for status which are notifications. 
### Testing

The test for the legacy implementation of Notifications are moved from `NotificationsConfigIT` to `LegacyNotificationsConfigIT` and they applied to Bolt Protocol versions from `5.4` to `5.2`. `NotificationConfigIT` is responsible to tests Bolt 5.5 or newer.

Unit tests were added to `DefaultMetadataHandler` to make sure all the data types present in the `status_parameters` as supported, at the moment `int`, `boolean`, `String` and `List<String>`. See, `CommonGqlStatusObjectImplementation::insertMessageParameters`.

Co-authored-by: grant lodge <6323995+thelonelyvulpes@users.noreply.github.com>
Co-authored-by: dotStart <dotstart@pm.me>