Remove pluggable DatabaseAllocator
There were two significant DatabaseAllocator implementations in the codebase, from which one was turned off. We don't see in the future that the DatabaseAllocator needs to be customized on this level, therefor the service provider pattern is unnecessary and also kind of messy. Removing the pattern means, that we can also:
* Remove the property form the topology config node
* Deprecate the `setDatabaseAllocator` procedure and make it into a no-op.
* Deprecate the `initial.dbms.database_allocator` setting.
* Remove the `OptimalDiskUtilizationAllocator`