test: Windows line safety via helper methods
Instead of spreading `WindowsSafeAnyRef` all over our tests with a comment explaining why it is needed, we can also introduce local, more targeted solutions like for solved strings for nested plan expressions as we do in this PR.

Introduced `LogicalPlanAstConstructionTestSupport` as an extension to `AstConstructionTestSupport`, which allows to construct line-break aware nested plan expressions. How this can simplify a test file, is shown on `SubqueryExpressionPlanningIntegrationTest` as an example.