SPD: disable parallel access check when beginning a diver transaction
The driver might go to System database when figuring out the cluster topology If this happens it must be performed with the parallel access check disabled in case the operation is performed from a parallel runtime worked thread. Parallel access check checks that Cypher worker threads don't wonder off and touch something they should not. This case is a false positive.