Makes listening to index updates similar to listening to ID updates
So that e.g. BlockTransactionApplier doesn't need to be bound to
accumulating index updates and having a work-sync managing the
index updates.

The basic change is to introduce an interface (IndexUpdatesListener)
which listens to the index updates, and that IndexUpdatesWorkSync.Batch
implements (very similar to ID updates). BlockTransactionApplier can
now work with any IndexUpdatesListener (plural).

The existing IndexUpdateListener (singular) name was kept to not
cause a larger change. This interface still has the same role
as before, i.e. used for creating/activating/dropping indexes
and applying batches of index updates, which IndexingService
is the main implementor of.