Last committed batch now correctly updated by concurrent transactions
in 5.23 a change was made which made appendBatch be called without synchronization. That meant that last committed batch could be updated out of order. This is now fixed by always keeping the highest batch.

The highest appendBatch is the only one of interest for lastCommittedBatch,
all batches of lower append indexes are guaranteed to have been appended to log already and there is no need to wait for their calls before publishing the highest one.

adds HighestAppendBatch and accompanying unit tests

Co-authored-by: Emma Holmberg Ohlsson <emma.holmberg.ohlsson@neo4j.com>