Don't fill ID caches (synchronously) on startup
Startup time of the dbms is important and so doing work that can
just as well be done later by either background threads or
transactional threads should be avoided if not necessary.
This commit removes the filling of ID caches during the start
of a db, since it could potentially take a non-negligible
amount of time. Note that there are two ways that ID caches gets
filled at runtime:
- the MAINTENANCE background thread(s) that runs at least
  once per second
- transactions noticing empty ID cache

So removing this fill during startup doesn't change any behaviour.