Rework checkConstraintsAndGetIndexes() logic
In order to avoid checking orphaned indexes (indexes that were supposed
to be attached to a constraint but the constraint does not exists), we
go from constraints -> indexes instead of the other way around.

This fixes a bug with the addition/removal of labels that are indexed by
an orphaned index.