Compact node ids by default when migrating to other storage engine
Bring back compaction of the node id space by default when migrating
to another storage engine. It was disabled because of benchmark migrations needing
to have the same ids to be able to be comparable. This introduces a hidden option to
toggle compaction instead that the benchmark migrations should use.