fix: Plan sort before aggregation in parallel runtime () ()
The COLLECT aggregation function should collect entries in the order determined by the previous ORDER-BY clause. This is broken in the parallel runtime if we're using any subqueries in the grouping/aggregating expressions, since operators like letSemiApply, semiApply, etc do not maintain order. Also, for sharded databases, if there's a remoteBatchProperties operator before an eagerAggregation, the COLLECT will no longer maintain order. Therefore, to preserve ORDER for certain aggregating functions like COLLECT, this PR will attempt to sort again, IN CASE the order has not been preserved. To facilitate this, operations like RemoteBatchingProperties and SubqueryExpressionSolvers have been moved out of the Aggregation method to their own. As a side-effect, an filters that are part of the aggregating projection now also get a remoteBatchProperties planned with them.

cl: Fixes a bug where in certain cases where the collect function wasn't preserving the order defined in the previous query clause in the parallel runtime.