test: Use line breaks from file not from System to format Cypher
Instead of slapping on another *ignore windows line break differences*, this change tries to eliminate differences by aligning the line breaks:

If on Windows, we can
- either have set `git config autocrlf` to `true` and get a file with `\r\n` line breaks, which means that multi-line strings will have that line break or
- have set `git config autocrlf` to `false` in which case `System.lineSeparator()` would still return `\r\n` but the multi-line strings would not. This created most of the conflicts that we tried to cover up.

The current change uses the kind of line breaks that are present in the file to format Cypher statements. This make sure that the expected results match the actual ones.

If this should be a problem, then we should re-think using multi-line strings.