Should be able to return startNode/endNode of deleted relationship
After merging, https://github.com/neo-technology/neo4j/pull/24677, we changed
the behavour of
```
> CREATE ()-[:R]->();
> MATCH (n)-[r:R]->() DETACH DELETE n RETURN startNode(r);
```
from returning a `-1` node to failing with the cryptic error message

```
TransientException: Database elements [...] were observed during query execution,
but got deleted by an overlapping committed transaction before the query results could be serialised.
The transaction might succeed if it is retried.
```
This PR changes so that we get the old, slightly better behavior.

changelog: Don't throw a transient when returning `startNode` and `endNode` on a deleted relationship.