Improve `SystemDatabaseProvider`
`SystemDatabaseProvider` has a `queryIfAvailable` method - which were only checking if the database was available, but did throw if the database context did not exist. This is not the best; `queryIfAvailable` should mean "give me an answer if you can", not "in not cases you can return an empty optional, but in some case we will throw". From the perspective of the caller there is no difference between the context not being there yet, or the database not being available yet.

Generally all methods returning `Optional<?>` should not throw an `SystemDatabaseUnavailableException`.