Correct max inline size for internal GBPTree nodes
Fixes a bug where trying to store keys close to the max inline size
could result in a GSPP write error (and fail applying transactions
already written to the log). The max inline size for internal tree nodes
was wrong and did not take the child pointers into consideration. This
meant that the offset array growing from the left and the keys growing
from the right could be competing for the same space and potentially
overlap. The write of the rightmost child pointer would fail when
reading the current state in its spot before the write - which would be
the beginning of an actual key. Keep in mind that the internal nodes
store splitter keys, not the actual keys, but the minimal key that
separates the two values on either side. This means that this would only
happen if two splitter keys of about 4000ish bytes (about 3996-4047 if
our calculations are correct) end up in the same internal node.

For existing indexes there is a chance that they can still run into GSPP
write errors if they already have a splitter key in an internal node of
a size between the new max inline size and the old one. If that key
would be combined with a new key of the new max inline size the keys and
offset array could overlap again. It is very unlikely and anyone running
into it should get the same GSPP write error. If that ever happens a
rebuild of the index will help since all values will then use the new
inline limit.