Issue with morsel reuse () ()
For queries not producing any results, EmptyResult will call `truncate`
on the morsel, when that morsel is later reused we will not end up
clearing all cached properties in some cases.

This might lead to a regression since we will call `Arrays.fill` with a
bigger range. At the moment I don't see a better way of doing it except
for perhaps only clearing the actual cached properties and not all refs,
I am not sure that would be faster though.

cherry-picks: 


[benchmarks](https://benchmarking.neo4j.today/?route=_dashboards_branch-form&personal-projects=pontusmelke/dev-morsel-reuse-bug&public-projects=dev&projects=pontusmelke/dev-morsel-reuse-bug&load-charts%3F=true)