Fix Modern Handshake for limited protocol versions ()
The modern handshake was offering all available protocol version in the
code base instead of using the registered ones.
This causes the connection close when the client selects an available
version which isn't registered due to configuration.

For example, Bolt 6.0 is available, but this version is registered by
default.

cherry-picks: