Fix deadlock in ClusterLeaseCoordinator + flaky tests ()
The flakiness in this test is related to a deadlock during
`ClusterLeaseCoordinator.acquireLeaseOrThrow()`.
We are trying to shutdown but if we're never trying to break the wait
while acquiring the lease, well...we could wait forever since here we
have only one leader with the other two servers that are shut down in
this test.

---------

Co-authored-by: grant lodge <6323995+thelonelyvulpes@users.noreply.github.com>