Clean transactions between QueryAPI tests ()
A series of tests have failed because of a single test failure. The
problem is that the transactions which are closed are kept from one test
to other, so the check for no-connections being open fails for the next
tests in the same class.

Cleaning the transactions between tests solves the state leakage between
tests.

Apart of this, the tests weren't testing if the transaction were
successfully committed. This makes the failures more difficult of being
captured.

Ensure that the commit was successfully done can help with debugging for
when a failure happens.