CVE-2025-11602 Fix init of self-allocated Bolt BitMask ()
The ByteBufAllocator usually used form the BitMask is pooled. In
scenarios where we different BitMasks are offered - with and without
FABRIC capability, for example - the newone can be created with data of
previous one.
In the example scenario, the FABRIC capability can be offered in a
connection which doesn't support it.

Erasing the buffer during the object creation solves the issue.

cherry-picks: 

---------

Co-authored-by: Yuki <dotstart@pm.me>