Enable heap estimator cache in top-level DISTINCT ()
When the grouping key for the seen set contains multiple values they are
wrapped in a list. We use a scoped heap estimator cache when
constructing the list to deduplicate heap estimations of large objects,
that could otherwise be cached as payload in potentially many new unique
list values.

Cherry-picks: