Fix dumped archive not being truncated with --overwrite-destination ()
…)

tldr; `Dumper` does not truncate archives when run with
`--overwrite-destination`, and Zstd throws you an 'unknown frame
descriptor' error for it.

In https://github.com/facebook/zstd/pull/3175, released in Zstd 1.5.4, a
new error was introduced when the stream decoder has finished decoding a
frame. If there are more bytes in the stream, then it assumes that this
must correspond to a new frame - if not, then it determines that the
archive is corrupt with `Unknown frame descriptor`.

Before this error was added, Zstd would quietly ignore that the stream
did not contain new necessary marker, `ZSTD_MAGICNUMBER`, or one of the
16 `ZSTD_MAGIC_SKIPPABLE` headers, happily keep consuming the contents.

When loading dumps, this can happen as the `TarArchiveInputStream`
reaches the `EOF` marker in the TAR archive (two 512 blocks of zeroes).
The stream will then attempt to skip to the end of the file.As it does
so, it pulls more data from the `ZstdInputStream`, which will attempt to
decode a new frame on the trash data.

In a correct archive, there should not be more data after the EOF
marker, but due to an unfortunate bug, our archives can contain more
data under a special circumstance, viz. when running the `Dumper` with
the `--override-destination` flag.

The root cause is found in the `DefaultFileSystemAbstraction` where the
`openAsOutputStream`-method does not truncate files when they are opened
with `append=false`, despite its documentation claiming that is should
do so. Instead it just opens them for write with file offset 0, meaning
that if the original dump was larger than its successor, the successsor
will have trash data trailing its valid Zstd frame.

As far as I can tell, this issue has been present in Neo4j since 5.19.
It seems that the issue was introduced in the refactoring done in
https://github.com/neo-technology/neo4j/pull/23520, which changed the
`Files.newOutputStream` (which does truncate) to `fs.openAsOutputStream`
(which does not).

Cherry-picks: